Misc fixes, mainly split wav writer and oss plugin and matrix fixes.

+++ SpiralSound/Plugins/OscillatorPlugin/OscillatorPluginGUI.C
@@ -328,7 +328,7 @@ void OscillatorPluginGUI::cb_Noise(Fl_Check_Button* o, void* v)
 inline void OscillatorPluginGUI::cb_SHLen_i(Fl_Slider* o, void* v)
 {
 	char str[10]; 
-	m_GUICH->Set("SHLen",0.2f-o->value()); 
+	m_GUICH->Set("SHLen",(float)(0.2f-o->value())); 
 	sprintf(str,"%4.3f s", 0.2f-o->value());
 	m_out_SHlen->value(str);
 }
@@ -339,7 +339,7 @@ void OscillatorPluginGUI::cb_SHLen(Fl_Slider* o, void* v)
 inline void OscillatorPluginGUI::cb_ModAmount_i(Fl_Knob* o, void* v)
 {
 	char str[10];
-	m_GUICH->Set("ModAmount",o->value());
+	m_GUICH->Set("ModAmount",(float)o->value());
 	sprintf(str,"%4.0f %%", 100*o->value());
 	m_out_mod->value(str); 
 }
